tests/submission/Infraware/Forms/contents/Forms/output_type.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_type</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <output name=o id='output_id'></output>
	</form>

	<script>
		
		var output = document.getElementById("output_id");
		type = output.type;

		if (typeof(output.type) == "string") {
			test(function() {
				assert_equals(output.type, "output", "type attribute is not exist.");
			});
		} else {
			test(function() {
				assert_unreached("type attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
